requirements.txt
evaluate
pandas
numpy
peft
PyYAML
tqdm
transformers
mlx
networkx 
openai 